Broke out the partition check from main.js, including dependent functions including countLiveTiles, countDeadTiles, and hasBeenCounted. #28
